requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-generativeai
pydantic[dotenv]
python-dotenv
sse-starlette
